requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pyyaml==6.0.2
rapidfuzz==3.9.6